% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scr.R
\name{avoidNetOverlaps}
\alias{avoidNetOverlaps}
\title{Avoid Overlaps in the Network}
\usage{
avoidNetOverlaps(w, threshold = 0.1)
}
\arguments{
\item{w}{A data frame with the x and y coordinates of the nodes.}

\item{threshold}{The threshold to avoid label overlaps.}
}
\value{
The labels to be removed.
}
\description{
This function is used to avoid label overlaps in the network.
 The function returns the labels to be removed.
}
\note{
This function was copied from Bibliometrix package, available at:
https://cran.r-project.org/web/packages/bibliometrix/bibliometrix.pdf
}
