% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{biblio_ds}
\alias{biblio_ds}
\title{Title: biblio_ds}
\usage{
biblio_ds(db_params)
}
\arguments{
\item{db_params}{A list of parameters for the database}

\item{db_params$raw_data_path}{The path to the raw data}

\item{db_params$out_path}{The path to the output folder}

\item{db_params$origin}{The origin of the data}

\item{db_params$format}{The format of the data}

\item{db_params$delim}{The delimiter/separator of the data}

\item{db_params$years}{The years to consider}

\item{db_params$n_max}{The maximum number of elements to consider}
}
\value{
A data frame with the data
}
\description{
It reads the data from the raw data path, converts it to a data frame, and
filters only the articles in English. It also removes duplicates from the data
frame. It modifies globally the db_params argument
}
