% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{collab_networks}
\alias{collab_networks}
\title{Collaborative Networks}
\usage{
collab_networks(M, db_params, default_params_network)
}
\arguments{
\item{M}{A data frame with the data}

\item{db_params}{A list of parameters for the database}

\item{default_params_network}{A list of parameters for the network}
}
\value{
Does not return anything, but writes the results of the analysis to
       the output folder
}
\description{
This function creates collaborative networks for the authors, universities,
and countries fields.
}
