% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{coocurrence_networks}
\alias{coocurrence_networks}
\title{Co-Ocurrence Networks}
\usage{
coocurrence_networks(M, db_params, default_params_network)
}
\arguments{
\item{M}{A data frame with the data}

\item{db_params}{A list of parameters for the database}

\item{default_params_network}{A list of parameters for the network}
}
\value{
Does not return anything, but writes the results of the analysis to
       the output folder
}
\description{
This function creates co-occurrence networks for the authors, keywords, and
author-keywords fields. It also tries to create networks for the titles,
abstracts, and sources fields, but it does not work.
}
