% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scr.R
\name{dist2df}
\alias{dist2df}
\title{Table Tag}
\usage{
dist2df(inDist)
}
\arguments{
\item{inDist}{A matrix with the data.}
}
\value{
A table with the data.
}
\description{
This function is used to create a table with the data.
}
\note{
This function was copied from Bibliometrix package, available at:
https://cran.r-project.org/web/packages/bibliometrix/bibliometrix.pdf
}
