% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explo_analys.R
\name{explorative_analysis}
\alias{explorative_analysis}
\title{Explorative Analysis}
\usage{
explorative_analysis(M, db_params, analysis_params)
}
\arguments{
\item{M}{A bibliometric database}

\item{db_params}{A list of parameters for the database}

\item{analysis_params}{A list of parameters for the analysis}
}
\value{
Does not return anything, but writes the results of the analysis to
        the output folder
}
\description{
This function performs an explorative analysis of the bibliometric database
provided. It includes the following analyses:
\itemize{
\item Summary of the data
\item Most cited articles and authors
\item Local citation analysis
\item Authors dominance
\item Authors h-index
}
}
\examples{
explorative_analysis(M, db_params, analysis_params)

}
