% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scr.R
\name{igraph2vis}
\alias{igraph2vis}
\title{Igraph To visNetwork}
\usage{
igraph2vis(
  g,
  curved = FALSE,
  labelsize = 3,
  opacity = 1,
  type = "auto",
  shape = "dot",
  net,
  shadow = TRUE,
  edgesize = 10,
  threshold = 0.05
)
}
\arguments{
\item{g}{A igraph network}

\item{curved}{Logical. If TRUE, the edges are curved.
If FALSE, the edges are straight.}

\item{labelsize}{The size of the labels.}

\item{opacity}{The opacity of the nodes.}

\item{type}{The type of layout to be used. Options are:
\itemize{
\item{"auto"}{Automatic layout}
\item{"circle"}{Circular layout}
\item{"mds"}{Multidimensional scaling layout}
\item{"star"}{Star layout}
\item{"sphere"}{Sphere layout}
\item{"fruchterman"}{Fruchterman layout}
\item{"kamada"}{Kamada layout}
}}

\item{shape}{The shape of the nodes.}

\item{net}{The network to plot.}

\item{shadow}{Logical. If TRUE, the nodes have shadow.}

\item{edgesize}{The size of the edges.}

\item{threshold}{The threshold to avoid label overlaps.}
}
\value{
The visNetwork network.
}
\description{
This function is used to convert a igraph network to a visNetwork network.
The function returns the visNetwork network.
}
\note{
This function was copied from Bibliometrix package, available at:
https://cran.r-project.org/web/packages/bibliometrix/bibliometrix.pdf
}
