% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scr.R
\name{netLayout}
\alias{netLayout}
\title{Network Layout}
\usage{
netLayout(type)
}
\arguments{
\item{type}{The type of layout to be used. Options are:
\itemize{
\item{"auto"}{Automatic layout}
\item{"circle"}{Circular layout}
\item{"mds"}{Multidimensional scaling layout}
\item{"star"}{Star layout}
\item{"sphere"}{Sphere layout}
\item{"fruchterman"}{Fruchterman layout}
\item{"kamada"}{Kamada layout}
}}
}
\value{
The layout to be used in the visNetwork function.
}
\description{
This function is used to define the layout of the network.
The layout can be defined as automatic, circular, multidimensional scaling,
star, sphere, Fruchterman, or Kamada.
The function returns the layout to be used in the visNetwork function.
}
\note{
This function was copied from Bibliometrix package, available at:
https://cran.r-project.org/web/packages/bibliometrix/bibliometrix.pdf
}
