% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_scr.R
\name{plot_net}
\alias{plot_net}
\title{Plot and save the network}
\usage{
plot_net(net, filename, db_params)
}
\arguments{
\item{net}{Network to plot.}

\item{filename}{Output filename.}

\item{db_params}{Parameters to save the file.}
}
\value{
Does not return anything, but saves the network as a PNG and a html file.
}
\description{
This function is used to plot and save the network. It uses webshot2
to take a picture of the network.
}
