% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{plot_network}
\alias{plot_network}
\title{Plot_network}
\usage{
plot_network(
  M,
  analysis,
  network,
  db_params = db_params,
  default_params_network = default_params_network
)
}
\arguments{
\item{M}{A data frame with the data}

\item{analysis}{The type of analysis that will be performed}

\item{network}{The type of network that will be plotted}

\item{db_params}{Parameters for the database}

\item{default_params_network}{Parameters for the network}
}
\value{
Does not return anything, but saves a PNG and a html, if specified
to the output folder
}
\description{
This function creates a network plot for the specified analysis and network
type. It also saves the plot as a PNG and a html file.
}
