% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_graphs.R
\name{summary_graphs}
\alias{summary_graphs}
\title{Summary Graphs}
\usage{
summary_graphs(M, db_params = db_params, analysis_params = analysis_params)
}
\arguments{
\item{M}{A data frame with the data}

\item{db_params}{Parameters for the database}

\item{analysis_params}{Parameters for the analysis}
}
\value{
Does not return anything, but writes the results of the analysis to
      the output folder
}
\description{
It creates a summary of the analysis based on the data frame M
and the parameters specified
}
