% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_analys.R
\name{word_cloud}
\alias{word_cloud}
\title{Word Cloud}
\usage{
word_cloud(M, field = "DE", n = 100, db_params)
}
\arguments{
\item{M}{A data frame with the data}

\item{field}{What field to use for the word cloud}

\item{n}{Number of words to include in the word cloud}

\item{db_params}{Parameters for the database}
}
\value{
Does not return anything, but writes the results of the analysis to
     the output folder
}
\description{
It creates a word cloud based on the data frame M and the field specified
}
